HolaBondi Project
HolaBondi is a program that provides the user, after 3 simple selections, the buses arrival time for the city of Cordoba. It covers 20 local bus lines. The main algorithm is a xls file parser, since the time-sheets provided by the bus companies are in that format. The program is designed in a way that has an easy customization. The second phase of the project is to convert it to an Android app that will run offline with no need of users to use the internet to get the estimated arrival times.
Hola Bondi was developed as final project of 2017 CS50 Harvard course by Gaston Mugas.